Implemented changing competition varibles, improved chart rendering
- Implemented changing of competition variables on CompSettings.vue. All variables are changeable, and after changing the competition is re-defined and results calculated

- Improved rendering of charts on Results.vue and PublicResults.vue.  Now the are responsive on all cases, and are fixed size when included in pdf

- Added addition button to Weighting.vue to add together multiple sums

+++ server/routes/api/cups.js
@@ -6,53 +6,85 @@ const router = express.Router();
 // Get all cups
 router.get("/", async (req, res) => {
   const cups = await loadCupsCollection();
-  res.send(await cups.find({}).toArray());
+  if (cups) {
+    res.send(await cups.find({}).toArray());
+  } else {
+    // Connection timed out
+    res.status(408).send();
+  }
 });
 
 // Get user's cups
 router.get("/:user_id", async (req, res) => {
   const cups = await loadCupsCollection();
-  res.send(await cups.find({ user_id: req.params.user_id }).toArray());
+  if (cups) {
+    res.send(await cups.find({ user_id: req.params.user_id }).toArray());
+  } else {
+    // Connection timed out
+    res.status(408).send();
+  }
 });
 
 // Get certain cup
 router.get("/cup/:cup_id", async (req, res) => {
   const cups = await loadCupsCollection();
-  res.send(
-    await cups.find({ _id: new mongodb.ObjectID(req.params.cup_id) }).toArray()
-  );
+  if (cups) {
+    res.send(
+      await cups
+        .find({ _id: new mongodb.ObjectID(req.params.cup_id) })
+        .toArray()
+    );
+  } else {
+    // Connection timed out
+    res.status(408).send();
+  }
 });
 
 // Add Cup
 router.post("/", async (req, res) => {
   const cups = await loadCupsCollection();
-  await cups.insertOne({
-    user_id: req.body.user_id,
-    name: req.body.name,
-    year: req.body.year,
-    signees: req.body.signees,
-    isPublic: req.body.isPublic,
-    createdAt: new Date(),
-  });
+  if (cups) {
+    await cups.insertOne({
+      user_id: req.body.user_id,
+      name: req.body.name,
+      year: req.body.year,
+      signees: req.body.signees,
+      isPublic: req.body.isPublic,
+      createdAt: new Date(),
+    });
 
-  res.status(201).send();
+    res.status(201).send();
+  } else {
+    // Connection timed out
+    res.status(408).send();
+  }
 });
 
 // Update one competition
 router.put("/:id", async (req, res) => {
   const cups = await loadCupsCollection();
-  const cup = req.body;
-  delete cup._id;
-  await cups.replaceOne({ _id: new mongodb.ObjectID(req.params.id) }, cup);
-  res.status(204).send();
+
+  if (cups) {
+    const cup = req.body;
+    delete cup._id;
+    await cups.replaceOne({ _id: new mongodb.ObjectID(req.params.id) }, cup);
+    res.status(204).send();
+  } else {
+    // Connection timed out
+    res.status(408).send();
+  }
 });
 
 // Delete Cup
 router.delete("/:id", async (req, res) => {
   const cups = await loadCupsCollection();
-  await cups.deleteOne({ _id: new mongodb.ObjectID(req.params.id) });
-
-  res.status(200).send();
+  if (cups) {
+    await cups.deleteOne({ _id: new mongodb.ObjectID(req.params.id) });
+    res.status(200).send();
+  } else {
+    // Connection timed out
+    res.status(408).send();
+  }
 });
 
 async function loadCupsCollection() {
@@ -64,12 +96,16 @@ async function loadCupsCollection() {
     mongodb_url = config.mongodb_url;
   }
 
-  const client = await mongodb.MongoClient.connect(mongodb_url, {
-    useNewUrlParser: true,
-    useUnifiedTopology: true,
-  });
+  try {
+    const client = await mongodb.MongoClient.connect(mongodb_url, {
+      useNewUrlParser: true,
+      useUnifiedTopology: true,
+    });
 
-  return client.db("fisustaja").collection("cups");
+    return client.db("fisustaja").collection("cups");
+  } catch (err) {
+    console.log(err.message);
+  }
 }
 
 module.exports = router;
