+++ server/routes/api/feedback.js
@@ -6,27 +6,44 @@ const router = express.Router();
 // Get Competitions
 router.get("/", async (req, res) => {
   const feedback = await loadFeedbackCollection();
-  res.send(await feedback.find({}).toArray());
+
+  if (feedback) {
+    res.send(await feedback.find({}).toArray());
+  } else {
+    // Connection timed out
+    res.status(408).send();
+  }
 });
 
 // Add Competition
 router.post("/", async (req, res) => {
   const feedback = await loadFeedbackCollection();
-  await feedback.insertOne({
-    type: req.body.type,
-    message: req.body.message,
-    createdAt: new Date(),
-  });
 
-  res.status(201).send();
+  if (feedback) {
+    await feedback.insertOne({
+      type: req.body.type,
+      message: req.body.message,
+      createdAt: new Date(),
+    });
+
+    res.status(201).send();
+  } else {
+    // Connection timed out
+    res.status(408).send();
+  }
 });
 
 // Delete Competitions
 router.delete("/:id", async (req, res) => {
   const feedback = await loadFeedbackCollection();
-  await feedback.deleteOne({ _id: new mongodb.ObjectID(req.params.id) });
 
-  res.status(200).send();
+  if (feedback) {
+    await feedback.deleteOne({ _id: new mongodb.ObjectID(req.params.id) });
+    res.status(200).send();
+  } else {
+    // Connection timed out
+    res.status(408).send();
+  }
 });
 
 async function loadFeedbackCollection() {
@@ -38,12 +55,16 @@ async function loadFeedbackCollection() {
     mongodb_url = config.mongodb_url;
   }
 
-  const client = await mongodb.MongoClient.connect(mongodb_url, {
-    useNewUrlParser: true,
-    useUnifiedTopology: true,
-  });
+  try {
+    const client = await mongodb.MongoClient.connect(mongodb_url, {
+      useNewUrlParser: true,
+      useUnifiedTopology: true,
+    });
 
-  return client.db("fisustaja").collection("feedback");
+    return client.db("fisustaja").collection("feedback");
+  } catch (err) {
+    console.log(err.message);
+  }
 }
 
 module.exports = router;
